+++ docs/src/release_notes.rst
@@ -10,6 +10,8 @@ Added
 - Support Python 3.9 (:issue:`219`)
 - Dependency to HTTPX upgraded to include versions ``0.15.*`` and ``0.16.*``
   (:issue:`216`)
+- Error messages for :func:`parse_code_from_url`, :func:`parse_state_from_url`
+  and :meth:`Credentials.pkce_user_authorisation` were improved (:issue:`218`)
 
 3.1.0 (2020-09-13)
 ------------------